refactor matching service
